Move parse_cd_entries to its own file
mod.rs should only contain module declarations and re-exports.
